implemented level three of the game

+++ StarWarsDungeon/StarWarsLibrary/Weapon.cs
@@ -9,11 +9,11 @@ namespace StarWarsLibrary
     public class Weapon
     {
         public string Name { get; set; }
-        public int DamageMultiplier { get; set; }
+        public double DamageMultiplier { get; set; }
         public int Cost { get; set; }
 
         //constructor
-        public Weapon(string name, int damageMultiplier, int cost)
+        public Weapon(string name, double damageMultiplier, int cost)
         {
             Name = name;
             DamageMultiplier = damageMultiplier;
